Change enthalpy initialisation method
For specified Td_bp and state keyword, the starting values
for enthalpy must be recalculated even with good starting values
available from a previously converged solution. This is due to the fact,
that the specification of Td_bp will result in a linear dependency in case
the starting value of enthalpy is in two-phase region from the calculation
before. To prevent this, the starting value for enthalpy is recalulated
in every solution mode.